bot/requirement.txt
pyTelegramBotAPI
load_dotenv